Move section name step into its own component
