+++ packages/shared-components/src/room-list/SectionCreationView/SectionCreationView.module.css
@@ -6,9 +6,5 @@
  */
 
 .container {
-    min-height: 346px;
-}
-
-.form {
-    width: 100%;
+    height: 346px;
 }
