+++ packages/shared-components/src/room-list/SectionCreationView/SectionFormView.module.css
@@ -5,5 +5,6 @@
  * Please see LICENSE files in the repository root for full details.
  */
 
-export * from "./SectionCreationView";
-export type * from "./types";
+.form {
+    width: 100%;
+}
